First pre-release for testing. Fixed some error handling and added loads of improvements over Screenshot Uploader. added proper licences.
