+++ src/main/java/com/darkeyedragon/imageutils/client/enums/UploadRequest.java
@@ -1,0 +1,7 @@
+package com.darkeyedragon.imageutils.client.enums;
+
+public class UploadRequest{
+    public enum Type{
+        MENU, IN_GAME, IN_GAME_MENU
+    }
+}
