+++ src/main/java/com/darkeyedragon/imageutils/client/events/GuiOptionsHook.java
@@ -1,6 +1,6 @@
 package com.darkeyedragon.imageutils.client.events;
 
-import com.darkeyedragon.imageutils.client.gui.LocalScreenshots;
+import com.darkeyedragon.imageutils.client.gui.GuiLocalScreenshots;
 import com.darkeyedragon.imageutils.client.gui.buttons.GuiButtonWithImage;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiOptions;
@@ -13,7 +13,6 @@ public class GuiOptionsHook{
     public void onInit(GuiScreenEvent.InitGuiEvent e){
        if(e.getGui() instanceof GuiOptions){
            int width = e.getGui().width;
-           int height = e.getGui().height;
                                                         //ID,        x,                 y,        buttonWidth,    buttonHeight,     ????        ????         ????        ResourceLocation
            e.getButtonList().add(new GuiButtonWithImage(5, width-40, 20, 20, 20));
            e.setButtonList(e.getButtonList());
@@ -23,7 +22,7 @@ public class GuiOptionsHook{
     public void onButtonClick(GuiScreenEvent.ActionPerformedEvent e){
         if(e.getGui() instanceof GuiOptions){
             if(e.getButton().id == 5){
-                Minecraft.getMinecraft().displayGuiScreen(new LocalScreenshots(e.getGui()));
+                Minecraft.getMinecraft().displayGuiScreen(new GuiLocalScreenshots(e.getGui()));
             }
         }
     }
