--- src/main/java/com/darkeyedragon/imageutils/client/gui/CustomGuiConfig.java
@@ -1,11 +1,0 @@
-package com.darkeyedragon.imageutils.client.gui;
-
-import net.minecraft.client.gui.GuiScreen;
-import net.minecraftforge.fml.client.config.GuiConfig;
-
-public class CustomGuiConfig extends GuiConfig{
-
-    public CustomGuiConfig(GuiScreen parentScreen, String modid, String title){
-        super(parentScreen, modid, title);
-    }
-}
