--- src/main/java/com/darkeyedragon/imageutils/client/gui/GuiUploaderSelector.java
@@ -1,7 +1,0 @@
-package com.darkeyedragon.imageutils.client.gui;
-
-import net.minecraft.client.gui.GuiScreen;
-
-public class GuiUploaderSelector extends GuiScreen{
-
-}
