+++ src/main/java/com/darkeyedragon/imageutils/client/message/Messages.java
@@ -16,22 +16,32 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.regex.Pattern;
 
+import static com.darkeyedragon.imageutils.client.utils.ImageUtil.addToLinkList;
+
 public class Messages{
+
+
     public static void uploadMessage(String result){
+
         ITextComponent uploadstr = new TextComponentTranslation("imageutil.message.upload.success").appendSibling(new TextComponentString(" "));
         ITextComponent linkText = new TextComponentString(result);
+        try{
+            ImageUtil.downloadFromUrl(new URL(result));
+        }catch(IOException e){
+            e.printStackTrace();
+        }
+        linkText.getStyle().setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new TextComponentTranslation("imageutil.message.upload.hover")));
         linkText.getStyle().setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, result));
         linkText.getStyle().setUnderlined(true);
         linkText.getStyle().setColor(TextFormatting.AQUA);
-        imageLink(result);
         Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(uploadstr.appendSibling(linkText));
     }
-    public static void errorMessage(int errorCode, String responseMessage){
+    public static void errorMessage(String errorMessage){
         ITextComponent errorText = new TextComponentTranslation("imageutil.message.upload.error");
         errorText.getStyle().setColor(TextFormatting.RED);
-        ITextComponent response = new TextComponentString(errorCode+": "+responseMessage);
+        ITextComponent response = new TextComponentString(errorMessage);
         response.getStyle().setColor(TextFormatting.RED);
-        ITextComponent errorMessage = new TextComponentTranslation("imageutil.message.upload.error1");
+        ITextComponent infoMessage = new TextComponentTranslation("imageutil.message.upload.error1");
         ITextComponent link = new TextComponentTranslation("imageutil.message.upload.errorlink");
         ITextComponent hover = new TextComponentString("https://github.com/DarkEyeDragon/ImageUtils");
         link.getStyle().setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://github.com/DarkEyeDragon/ImageUtils"));
@@ -40,7 +50,7 @@ public class Messages{
         link.getStyle().setHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, hover));
         Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(errorText);
         Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(response);
-        Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(errorMessage.appendSibling(new TextComponentString(" ")).appendSibling(link));
+        Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(infoMessage.appendSibling(new TextComponentString(" ")).appendSibling(link));
     }
     public static void imageLink(String link, String unformattedText){
         ImageUtilsMain.fixedThreadPool.submit(()->{
@@ -48,7 +58,7 @@ public class Messages{
                 try{
                     URL url = new URL(link);
                     BufferedImage downloadedImage = ImageUtil.downloadFromUrl(url);
-                    addToList(link, downloadedImage);
+                    addToLinkList(link, downloadedImage);
                     //TODO CLEAN UP
                     ITextComponent textLink = new TextComponentString("view image");
                     textLink.getStyle().setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, link));
@@ -77,7 +87,7 @@ public class Messages{
                 try{
                     URL url = new URL(link);
                     BufferedImage downloadedImage = ImageUtil.downloadFromUrl(url);
-                    addToList(link, downloadedImage);
+                    addToLinkList(link, downloadedImage);
                     //TODO CLEAN UP
                     ITextComponent textLink = new TextComponentString("view image");
                     textLink.getStyle().setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, link));
@@ -90,11 +100,6 @@ public class Messages{
                     Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(new TextComponentString(e.getMessage()));
                 }
             }
-        });
-    }
-    private static synchronized void addToList(String urlString, BufferedImage downloadedImage){
-        Minecraft.getMinecraft().addScheduledTask(() -> {
-            ImageUtilsMain.validLinks.put(urlString, downloadedImage);
         });
     }
 }
